requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
argon2-cffi==23.1.0
gunicorn==23.0.0
flask-htmx==0.3.1
Werkzeug==2.3.7
Jinja2==3.1.6
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
python-gnupg==0.5.0
boto3==1.28.38
